Completed the initial redesign of vertex arrays.
The new functionality defers most of the initialisation to the compiler,
leaving only the type information and the data collection to the user
(neither of these can be inferred by the compiler anyway).

There are two types of array:

1. vertex_array_buffer
2. vertex_element_buffer

vertex_array_buffers are used when all the vertex data, including
colours, normals, etc., are provided. The interface for a
vertex_array_buffer resembles a tuple, and it takes a std_layout_tuple
as its sole constructor argument.

vertex_element_buffers are a refinement over vertex_array_buffers, and
also take a sequence of integers that denote the occurrence of each
vertex.

Neither type requires binding before drawing anymore: in fact, this
actually causes a segmentation fault.

Work is still necessary. The segfault should be handled with more
gracefully (literally anything short of some other UB is better than a
segfault), and there's definitely some cleaning up that needs doing.