Vertex array (#58)
* Added standard-layout tuple.

All types are required to be standard-layout. std_layout_tuple is non-IS
conforming, and should not be used, except when communication with GPUs
is required (i.e. prefer std::tuple over doge::std_layout_tuple).

*Some* structured binding expansions are supported. There is currently
  a bug that prevents `auto [a, b] = t` from working, but it is
  perfectly okay to use `auto&& [a, b] = t`. This appears to have the
  same effect, and is good enough for now.

* Added Peter Sommerlad's unique_resource, and various CMake fixes

* Added a way to detect the enum value for a type supported by OpenGL.

* Removed forwards compatibility flag, as it is apparently not useful.

Source: https://stackoverflow.com/questions/13403807/glvertexattribpointer-raising-gl-invalid-operation

* Added metaprogramming sub-library.

* Completed the initial redesign of vertex arrays.

The new functionality defers most of the initialisation to the compiler,
leaving only the type information and the data collection to the user
(neither of these can be inferred by the compiler anyway).

There are two types of array:

1. vertex_array_buffer
2. vertex_element_buffer

vertex_array_buffers are used when all the vertex data, including
colours, normals, etc., are provided. The interface for a
vertex_array_buffer resembles a tuple, and it takes a std_layout_tuple
as its sole constructor argument.

vertex_element_buffers are a refinement over vertex_array_buffers, and
also take a sequence of integers that denote the occurrence of each
vertex.

Neither type requires binding before drawing anymore: in fact, this
actually causes a segmentation fault.

Work is still necessary. The segfault should be handled with more
gracefully (literally anything short of some other UB is better than a
segfault), and there's definitely some cleaning up that needs doing.